uploading grants.sql
I can't run it but I can't tell what's wrong with it. I had to change the name to grants because grant is an sql command. I used checks to do non negative on the fields that were non negative. Will be updating if I can't get it to work on my machine.